+++ lib/scala/distribution-manager/src/main/scala/org/enso/distribution/DistributionManager.scala
@@ -89,7 +89,19 @@ case class DistributionPaths(
     * precedence.
     */
   def editionSearchPaths: Seq[Path] =
-    customEditions ++ Seq(cachedEditions)
+    customEditions ++ Seq(cachedEditions) ++
+    bundledWithEngines(DistributionManager.EDITIONS_DIRECTORY)
+
+  /** Returns a sequence of paths to some subdirectory in all installed engines.
+    */
+  def bundledWithEngines(subdirectory: String): Seq[Path] = {
+    for {
+      enginesDir <- engineSearchPaths
+      enginePath <- FileSystem.listDirectory(enginesDir)
+      candidate = enginePath.toAbsolutePath.normalize / subdirectory
+      if Files.exists(candidate)
+    } yield candidate
+  }
 }
 
 /** Paths to secondary directories for additionally bundled engine
@@ -158,13 +170,14 @@ class DistributionManager(val env: Environment) {
     def canonizeSeq(paths: Seq[Path]): String =
       paths.map(canonize).mkString(File.pathSeparator)
     Map(
-      ENSO_DATA_DIRECTORY    -> canonize(paths.dataRoot),
-      ENSO_CONFIG_DIRECTORY  -> canonize(paths.config),
-      ENSO_RUNTIME_DIRECTORY -> canonize(paths.runRoot),
-      ENSO_LOG_DIRECTORY     -> canonize(paths.logs),
-      ENSO_HOME              -> canonize(paths.ensoHome),
-      ENSO_EDITION_PATH      -> canonizeSeq(paths.customEditions),
-      ENSO_LIBRARY_PATH      -> canonizeSeq(paths.localLibrariesSearchPaths)
+      ENSO_DATA_DIRECTORY           -> canonize(paths.dataRoot),
+      ENSO_CONFIG_DIRECTORY         -> canonize(paths.config),
+      ENSO_RUNTIME_DIRECTORY        -> canonize(paths.runRoot),
+      ENSO_LOG_DIRECTORY            -> canonize(paths.logs),
+      ENSO_HOME                     -> canonize(paths.ensoHome),
+      ENSO_EDITION_PATH             -> canonizeSeq(paths.customEditions),
+      ENSO_LIBRARY_PATH             -> canonizeSeq(paths.localLibrariesSearchPaths),
+      ENSO_AUXILIARY_LIBRARY_CACHES -> canonizeSeq(auxiliaryLibraryCaches())
     )
   }
 
@@ -184,8 +197,13 @@ class DistributionManager(val env: Environment) {
     * These locations can be used to preload published libraries, for example
     * from a shared network drive, so that they do not need to be downloaded.
     */
-  def auxiliaryLibraryCaches(): Seq[Path] =
-    env.getEnvPaths(ENSO_AUXILIARY_LIBRARY_CACHES).getOrElse(Seq())
+  def auxiliaryLibraryCaches(): Seq[Path] = {
+    val fromEnv =
+      env.getEnvPaths(ENSO_AUXILIARY_LIBRARY_CACHES).getOrElse(Seq())
+    val fromBundles =
+      paths.bundledWithEngines(DistributionManager.LIBRARIES_DIRECTORY)
+    fromEnv ++ fromBundles
+  }
 
   /** Finds the path to the ENSO_HOME directory that is used for keeping user's
     * projects, libraries and other custom artifacts.